pyproject.toml
[project]
name = "data-annotation-course"
version = "1.0.0"
description = "数据标注课程自动化解题脚本"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "requests",
    "Pillow",
    "opencv-python-headless",
    "easyocr",
    "transformers>=4.30.0",
    "torch",
    "tensorflow>=2.10.0,<2.22",
    "numpy",
    "scikit-image",
    "matplotlib",
]

[project.scripts]
data-anno = "scripts.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
python-version = "3.12"
python-preference = "only-system"
